requirements.txt
ultralytics>=8.0.0,<9.0.0
opencv-python==4.8.1.78
numpy==1.24.4
pandas==2.1.2
scikit-learn==1.2.2
matplotlib==3.8.0
seaborn==0.12.2
tqdm==4.66.1
pillow>=10.1.0
scipy>=1.11.4
torch==2.1.1
torchvision==0.16.1
scikit-image>=0.26.0
streamlit>=1.26.0
reportlab>=4.0
albumentations>=1.3.0

